backend/requirements.txt
# Backend Requirements for AI News Summary System
# Using DeepSeek API for summarization

# Web Framework
flask==3.0.0
flask-cors==4.0.0

# RSS Parsing
feedparser==6.0.11

# HTTP Requests
requests==2.31.0

# Note: Using direct requests instead of OpenAI SDK for better compatibility

# Environment Variables
python-dotenv==1.0.0

# Data handling
python-dateutil==2.8.2

# Caching (optional but recommended)
diskcache==5.6.3

# Scheduling (untuk auto-update summary)
schedule==1.2.1

